Set package list based on ROS distro
This change makes the build_docs.sh script backwards compatible with Dashing and Eloquent.
Since not all packages are available in all distros, we should only attempt to generate docs for those that exist.

This change requires us to explicitly name the packages that we are generated docs for in the Makefile.

Signed-off-by: Jacob Perron <jacob@openrobotics.org>